aml_hdmicec: use bin and lib to publish aml_hdmicec [1/2]
PD#SWPL-57427

Problem:
new feature.

Solution:
use bin and lib to publish aml_hdmicec.

Verify:
yocto ap222

+++ recipes-multimedia/aml-hdmicec/aml-hdmicec_git.bb
@@ -13,6 +13,8 @@ do_configure[noexec] = "1"
 do_compile[noexec] = "1"
 inherit autotools pkgconfig systemd
 S="${WORKDIR}/git"
+ARM_TARGET = "32"
+ARM_TARGET_aarch64 = "64"
 
 EXTRA_OEMAKE="STAGING_DIR=${STAGING_DIR_TARGET} \
                 TARGET_DIR=${D} \
@@ -22,10 +24,12 @@ do_install() {
     install -d ${D}${libdir}
     install -d ${D}${bindir}
     install -d ${D}${includedir}
-    install -m 0755 ${S}/hdmi-control-service ${D}${bindir}
 
-    if [ "${@bb.utils.contains("DISTRO_FEATURES", "systemd", "yes", "no", d)}" = "yes"  ]; then
-        install -D -m 0644 ${WORKDIR}/hdmicontrol.service ${D}${systemd_unitdir}/system/hdmicontrol.service
+    if [ -f "${S}/${ARM_TARGET}/hdmi-control-service" ]; then
+        install -m 0755 ${S}/${ARM_TARGET}/hdmi-control-service ${D}${bindir}
+        if [ "${@bb.utils.contains("DISTRO_FEATURES", "systemd", "yes", "no", d)}" = "yes"  ]; then
+            install -D -m 0644 ${WORKDIR}/hdmicontrol.service ${D}${systemd_unitdir}/system/hdmicontrol.service
+        fi
     fi
 }
 SYSTEMD_SERVICE_${PN} = "hdmicontrol.service "
